Notepad
Hello! This is my website for the exercise, I would really appreciate it if you read the readme completely. There I tell you about the requirements, the drawbacks and the development of the site. Without further ado, thank you very much, I hope you like the idea of ​​the project.